webapp/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
pydantic==2.6.4
aiofiles==23.2.1
xgboost==2.0.3
numpy==1.26.4
fair-esm==2.0.0
torch==2.2.2
biopython==1.83
pandas==2.2.1
